refactor: move all types to $lib/types for better management
