+++ src/lib/repos/events.ts
@@ -2,96 +2,7 @@ import { getDb } from '$lib/db';
 import { desc, eq, count, gt, like, or, and, asc } from 'drizzle-orm';
 import { events, attendees } from '$lib/db/schema';
 import { Logger } from '$lib/utils/logger';
-
-/**
- * Represents an event entity.
- */
-export type Event = {
-	id: string;
-	title: string;
-	slug: string;
-	description: string | null;
-	dateOfEvent: Date;
-	closeRsvpAt: Date | null;
-	maxAttendees: number | null;
-	image: string | null;
-	organizationId: string;
-};
-
-/**
- * Data required to create a new event.
- * @property title - Event title (required)
- * @property slug - Unique slug identifier (required)
- * @property description - Optional event description
- * @property dateOfEvent - Event date and time (required)
- * @property closeRsvpAt - Optional RSVP deadline
- * @property maxAttendees - Optional maximum attendee limit
- * @property image - Optional event image URL
- * @property organizationId - Organization hosting the event (required)
- */
-export type EventCreateData = {
-	title: string;
-	slug: string;
-	description?: string;
-	dateOfEvent: Date;
-	closeRsvpAt?: Date;
-	maxAttendees?: number;
-	image?: string;
-	organizationId: string;
-};
-
-/**
- * Data for updating an event. All fields optional except organizationId.
- */
-export type EventUpdateData = Partial<Omit<EventCreateData, 'organizationId'>>;
-
-/**
- * Represents an event attendee relationship.
- */
-export type Attendee = {
-	eventId: string;
-	userId: string;
-};
-
-/**
- * Represents an event organizer relationship.
- */
-export type EventOrganizer = {
-	eventId: string;
-	userId: string;
-};
-
-/**
- * Event with aggregated attendee count.
- */
-export type EventWithAttendeeCount = Event & {
-	attendeeCount: number;
-};
-
-/**
- * Pagination options for listing events.
- * @property page - Page number (1-based)
- * @property pageSize - Number of items per page
- */
-export type PaginationOptions = {
-	page: number;
-	pageSize: number;
-};
-
-/**
- * Result of a paginated query.
- * @property data - Array of results
- * @property pagination - Pagination metadata
- */
-export type PaginationResult<T> = {
-	data: T[];
-	pagination: {
-		page: number;
-		pageSize: number;
-		totalCount: number;
-		totalPages: number;
-	};
-};
+import type { Event, EventCreateData, EventOrganizer, EventUpdateData, EventWithAttendeeCount, PaginationOptions, PaginationResult } from "$lib/types"
 
 /**
  * Repository for event CRUD operations and queries.
