+++ tests/unit/events.crud.test.ts
@@ -1,11 +1,11 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import {
-	EventRepo,
 	type Event,
 	type EventCreateData,
 	type EventUpdateData,
 	type PaginationOptions
-} from '$lib/repos/events';
+} from '$lib/types';
+import { EventRepo } from '$lib/repos/events';
 
 vi.mock('$lib/db', () => ({
 	getDb: vi.fn(),
