+++ tests/unit/orgs.crud.test.ts
@@ -1,11 +1,11 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import {
-	OrganizationRepo,
 	type Organization,
 	type OrganizationCreateData,
 	type OrganizationUpdateData,
-	type PaginationOptions
-} from '$lib/repos/orgs';
+	type PaginationOptions,
+} from '$lib/types';
+import { OrganizationRepo } from '$lib/repos/orgs';
 
 vi.mock('$lib/db', () => ({
 	getDb: vi.fn(),
