+++ tests/unit/orgs.members.test.ts
@@ -1,5 +1,8 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { OrganizationRepo, type Organization } from '$lib/repos/orgs';
+import {
+	type Organization,
+} from '$lib/types';
+import { OrganizationRepo } from '$lib/repos/orgs';
 
 vi.mock('$lib/db', () => ({
 	getDb: vi.fn(),
