+++ tests/unit/orgs.queries.test.ts
@@ -1,5 +1,9 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { OrganizationRepo, type Organization, type OrganizationWithStats } from '$lib/repos/orgs';
+import {
+	type Organization,
+	type OrganizationWithStats,
+} from '$lib/types';
+import { OrganizationRepo } from '$lib/repos/orgs';
 
 vi.mock('$lib/db', () => ({
 	getDb: vi.fn(),
